requirements.txt
# Core Flask Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.10.0
Werkzeug==2.3.7

# Database (pre-compiled binary - most reliable)
psycopg2-binary==2.9.10

# Authentication & HTTP
Authlib==1.2.1
requests==2.31.0

# Machine Learning (versions with guaranteed wheel support)
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
xgboost==2.0.3
lightgbm==4.1.0

# Google Services
google-generativeai==0.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# Payment Gateway
razorpay==1.4.2

# Utilities
python-dotenv==1.0.0
Pillow==10.4.0
python-dateutil==2.8.2

# PDF Generation
reportlab==4.0.4
matplotlib==3.7.2

# Production Server
gunicorn==21.2.0